Import foreign model loadings

`make_loading` method takes the loading matrix obtained by a mixOmixs package and fold them into two-dimensional matrix

Usage

make_loadings(floadings, time, mod_time, acq_rate)

Arguments

floadings

a numeric matrix with foreign loadings. With variables in columns and eigenvalues as rows.

time

a vector of length two with the time range of the chromatographic run

mod_time

the modulation time of the second dimension.

acq_rate

the acquisition rate of the mass analyzer. The acquisition rate is printed when read_chrom function is performed

Details

We strongly recommend to use the plsda function in the mixOmics package to perform partial least squares-discriminant analysis. The result of this model is a list containing a loading matrix. The method retrieves a matrix A with m and n dimensions. Where m is the eigenvalues and n is the number of loadings which the model returns.